fix LVM in containers
The basic idea is:
- run --privileged and -v /dev:/dev, so that the container and host see
the same /dev file system.
- disable udev as much as possible in the container.
- It is the host's udev job to create for example the
/dev/poolname/lvname symlinks.

Unfortunately Debian's LVM has an unfixed bug that always tries to ask
udev. This was reported (#932433), but so far even unanswered. Fix it
for now with a custom package. This is a plain rebuild with the patch
mentioned in the Debian bug report.