repos: switch to LINBIT's piraeus repos
The Python code was switched to Python3, which broke the last release.
This is because Buster (the base for the Piraeus images) and Bionic
(from where we took the packages) have different Python 3 versions.
Don't break Debian, I should have known better.

As part of its Piraeus support LINBIT now provides a dedicated public
repository for Piraeus which contains a small subset of the packages
LINBIT provides to its customers. This is Debian Buster based.