piraeus-server: allow to use host LVM
LVM is the worst piece of software ever written. Depending on the host
system one sees pvs/vgs on the host. Or not. In our particular case vgs
show up on Centos8 and Bionic, but don't on Centos7 and Xenial.

This allows one to bind /proc to /host/proc and therefore trigger some
namespace trickery to actually use the hosts LVM.

"super privileged containers":
https://developers.redhat.com/blog/2014/11/06/introducing-a-super-privileged-container-concept/