add k8s-await-election helper to piraeus server
This entrypoint enables multiple replicas for LINSTOR controllers.
The basic principle is that only one controller is running at any
point in time. k8s leader election is used to coordinate which pod
gets to start the controller process.

If a controller dies or loses leader status another way, the process
is killed and another pod get to start the controller process.