piraeus-server: make controller log group-writable
In some Kubernetes contexts (OpenShift) we are not allowed to run as
root by default. This means that the controller may not have the required
permissions to create the log directoy.

This change manually creates the controller log directory and makes it
group-writable. Since we can ensure that the container user is a member of
this group (even as unprivileged SA in Openshift), this allows the controller
to write logs on all platforms.

Note that this is not an issue for satellites, as they always start as
privileged containers, i.e. they run as UID 0.