piraeus-server: use k8s-await-election v0.2.3
This version of k8s-await-election can set the "nodeName" attribute
on created endpoint resources. This enables Kubernetes "hairpin mode",
which is required for a pod to access it's own service.

This allows "linstor ..." commands from the controller container to succeed.
Previously they would never receive a response from the controller.