Ensure LINSTOR components log to files
Enable logging to files for LINSTOR controller and satellite. The default
(only logging to stdout) meant that sos-reports would not contain any logs.

The default file backend that is configured but not enabled uses log rotation
(keep up to 10 log artifacts, each limited to 2MB), which should be a good
default for most environments.