docker: disable LVM backup
By default, LVM tools will create a "metadata backup" every time a pool
is modified. Since the backup location only exists while the container
exists, this wasn't very useful in the container context.

In the interest of making our images ready to be read-only, disable the
creation of LVM backups.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>