docker: remove dmeventd dependency from linstor
When running in a container, we cannot assume to have access to dmeventd.
Up to now, we manually started dmeventd to work around an issue when running
the container on systemd with high rlimits. See 2b2ec174bcde for details.

However, there is no need to start dmeventd at all: it's primary use for
automatic resizing VGs and LVs doesn't fit the use in containers anyways.
So we can just disable "monitoring" of LVs, which in turn should prevent
dmeventd from being started by some random lvcreate.

Also, use "exec" to start the linstor processes: that way they get to be
PID 1 in the container and they receive any SIGINTs generated by the
container runtime.

Signed-off-by: Moritz Wanzenböck <moritz.wanzenboeck@linbit.com>