docker: use "tini" when starting tlshd
If a normal C program is started as PID 1, it will not have the default
SIGTERM handler installed: instead the signal will be ignored. This means
that during container shutdown, the container will run until forceful
shutdown via SIGKILL.

To work around this issue, we use "tini", a tiny init system for containers
to act as PID 1 in the container. This way, tlshd is started as PID 2 and gets
the normal SIGTERM handling, as tini will forward these signals.

This is only needed for ktls-utils:
* drbd-reactor has custom SIGTERM handling.
* LINSTOR inherits SIGTERM handling via the JVM.

Signed-off-by: Moritz Wanzenböck <moritz.wanzenboeck@linbit.com>