docker: use backports for zfsutils
The userspace part of zfs seems to have a major breakage between 2.1 and 2.2
releases. This means that userspace utilities <= 2.1 cannot interact in a
meaningful way with a 2.2 ZFS kmod.

This is causing issues with users that already have ZFS 2.2, which may get even
more severe in the future, as Ubuntu 24.04 will ship with ZFS 2.2 out of the
box.

To fix the issue, install the zfs userspace utilities from debian backports:
they seem to be reasonably recent.

Tested with versions from ZFS 2.2 to 0.7.

Signed-off-by: Moritz Wanzenböck <moritz.wanzenboeck@linbit.com>