ci: ensure separate caches
By default, all images get cached to the same "scope" in github. This means
that while we build the images concurrently, they all compete for the same
cache location.

This means that our cache only applies to (at most) one image, making it
not that useful. To solve this, we have to provide separate "scopes" for
all images.

Signed-off-by: Moritz Wanzenböck <moritz.wanzenboeck@linbit.com>