docker: use common base image
In order to configure the LINBIT repositories, we need to install the
linbit-keyring package. This requires some set up as we need to fetch the
package and then set up the repositories.

To make this nicer to work with, we use a single base docker image as the root
of all builds.

While at it, also make use of proper tmpfs and cache mounts, so we do not add
unnecessary files into the image layers.

Signed-off-by: Moritz Wanzenböck <moritz.wanzenboeck@linbit.com>